carve out UDFCustomErrorTest as its own class
