+++ src/test/java/org/sqlite/UDFCustomErrorTest.java
@@ -1,0 +1,50 @@
+package org.sqlite;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.LinkedList;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * Tests User Defined Functions.
+ */
+public class UDFCustomErrorTest {
+    private Connection conn;
+    private Statement stat;
+
+    @BeforeEach
+    public void connect() throws Exception {
+        conn = DriverManager.getConnection("jdbc:sqlite:");
+        stat = conn.createStatement();
+    }
+
+    @AfterEach
+    public void close() throws SQLException {
+        stat.close();
+        conn.close();
+    }
+
+    @Test
+    public void customErr() throws SQLException {
+        Function.create(conn, "f9", new Function() {
+            @Override
+            public void xFunc() throws SQLException {
+                throw new SQLException("myErr");
+            }
+        });
+        assertThrows(SQLException.class, () -> stat.executeQuery("select f9();"));
+    }
+}
